# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация StudX API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Локальный сервер"
        }
    ],
    "paths": {
...
        "/courses_of_user": {
            "get": {
                "tags": [
                    "courses_of_user"
                ],
                "summary": "Получение всех записей о курсах пользователей",
                "operationId": "56c30ccb270a66533de991a9ce5a8bba",
                "responses": {
                    "200": {
                        "description": "Список записей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/сourses_of_user"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "courses_of_user"
                ],
                "summary": "Создание записи о курсе пользователя",
                "operationId": "746aa265f1965209a5160060cbed6fcc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "course_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "Идентификатор пользователя",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "course_id": {
                                        "description": "Идентификатор курса",
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "custom_course_name": {
                                        "description": "Пользовательское название курса",
                                        "type": "string",
                                        "example": "Мой курс",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Запись создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Запись создана"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/сourses_of_user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/courses_of_user/{id}": {
            "get": {
                "tags": [
                    "courses_of_user"
                ],
                "summary": "Получение записи о курсе пользователя",
                "operationId": "eb155abdf7cd472403fd96670a30a4ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор записи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Запись найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/сourses_of_user"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Запись не найдена"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "courses_of_user"
                ],
                "summary": "Удаление записи о курсе пользователя",
                "operationId": "308241aa2863b9c12b8873a38c7b7518",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор записи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Запись удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Запись удалена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Запись не найдена"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "courses_of_user"
                ],
                "summary": "Обновление записи о курсе пользователя",
                "operationId": "43dbca059c97eaa301ff636ed09aac2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор записи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "custom_course_name": {
                                        "description": "Пользовательское название курса",
                                        "type": "string",
                                        "example": "Обновленное название курса",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запись обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Запись обновлена"
                                        },
                                        "custom_course_name": {
                                            "type": "string",
                                            "example": "Обновленное название курса"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Запись не найдена"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "сourses_of_user": {
                "title": "Курс пользователя",
                "description": "Схема записи пользователя и курса",
                "required": [
                    "id",
                    "user_id",
                    "course_id"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор записи",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer",
                        "example": 1
                    },
                    "course_id": {
                        "description": "Идентификатор курса",
                        "type": "integer",
                        "example": 6
                    },
                    "custom_course_name": {
                        "description": "Пользовательское название курса",
                        "type": "string",
                        "example": "Мой курс",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Дата и время создания записи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-19T12:34:56Z"
                    },
                    "updated_at": {
                        "description": "Дата и время последнего обновления записи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-19T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}